processor: move runProcessor to processor.go
