+++ cdc/processor.go
@@ -698,3 +698,62 @@ func (p *processor) register(ctx context.Context) error {
 func (p *processor) deregister(ctx context.Context) error {
 	return p.etcdCli.DeleteProcessorInfo(ctx, p.captureID, p.id)
 }
+
+// runProcessor creates a new processor then starts it.
+func runProcessor(
+	ctx context.Context,
+	session *concurrency.Session,
+	info model.ChangeFeedInfo,
+	changefeedID string,
+	captureID string,
+	checkpointTs uint64,
+) (*processor, error) {
+	opts := make(map[string]string, len(info.Opts)+2)
+	for k, v := range info.Opts {
+		opts[k] = v
+	}
+	opts[sink.OptChangefeedID] = changefeedID
+	opts[sink.OptCaptureID] = captureID
+	filter, err := util.NewFilter(info.GetConfig())
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	sink, err := sink.NewSink(info.SinkURI, filter, opts)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	ctx, cancel := context.WithCancel(ctx)
+	errCh := make(chan error, 1)
+	go func() {
+		if err := sink.Run(ctx); errors.Cause(err) != context.Canceled {
+			errCh <- err
+		}
+	}()
+	processor, err := newProcessor(ctx, session, info, sink, changefeedID, captureID, checkpointTs)
+	if err != nil {
+		cancel()
+		return nil, err
+	}
+	log.Info("start to run processor", zap.String("changefeed id", changefeedID))
+
+	processor.Run(ctx, errCh)
+
+	go func() {
+		err := <-errCh
+		if errors.Cause(err) != context.Canceled {
+			log.Error("error on running processor",
+				zap.String("captureid", captureID),
+				zap.String("changefeedid", changefeedID),
+				zap.String("processorid", processor.id),
+				zap.Error(err))
+		} else {
+			log.Info("processor exited",
+				zap.String("captureid", captureID),
+				zap.String("changefeedid", changefeedID),
+				zap.String("processorid", processor.id))
+		}
+		cancel()
+	}()
+
+	return processor, nil
+}
