--- cdc/scheduler_test.go
@@ -1,54 +1,0 @@
-// Copyright 2019 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package cdc
-
-import (
-	"context"
-	"net/url"
-
-	"github.com/pingcap/check"
-	"github.com/pingcap/ticdc/pkg/etcd"
-	"github.com/pingcap/ticdc/pkg/util"
-	"go.etcd.io/etcd/embed"
-	"golang.org/x/sync/errgroup"
-)
-
-type schedulerSuite struct {
-	etcd      *embed.Etcd
-	clientURL *url.URL
-	ctx       context.Context
-	cancel    context.CancelFunc
-	errg      *errgroup.Group
-}
-
-var _ = check.Suite(&schedulerSuite{})
-
-// Set up a embed etcd using free ports.
-func (s *schedulerSuite) SetUpTest(c *check.C) {
-	dir := c.MkDir()
-	var err error
-	s.clientURL, s.etcd, err = etcd.SetupEmbedEtcd(dir)
-	c.Assert(err, check.IsNil)
-	s.ctx, s.cancel = context.WithCancel(context.Background())
-	s.errg = util.HandleErrWithErrGroup(s.ctx, s.etcd.Err(), func(e error) { c.Log(e) })
-}
-
-func (s *schedulerSuite) TearDownTest(c *check.C) {
-	s.etcd.Close()
-	s.cancel()
-	err := s.errg.Wait()
-	if err != nil {
-		c.Errorf("Error group error: %s", err)
-	}
-}
